+++ src/blob/header.rs
@@ -28,8 +28,10 @@ impl Header {
         let buf = file
             .read_exact_at_allocate(size as usize, 0)
             .await
+            .map_err(|err| err.into_bincode_if_unexpected_eof())
             .with_context(|| format!("failed to read from file: {:?}", file_name))?;
         let header: Self = deserialize(&buf)
+            .map_err(|err| Error::from(err))
             .with_context(|| format!("failed to deserialize header from file: {:?}", file_name))?;
         header.validate().context("header validation failed")?;
         Ok(header)