+++ CHANGELOG.md
@@ -4,6 +4,20 @@ Pearl changelog
 
 ## [Unreleased]
 #### Added
+
+
+#### Changed
+
+
+#### Fixed
+
+
+#### Updated
+
+
+
+## [0.19.0] - 2023-05-15
+#### Added
 - Use InMemoryIndex as an additional filter on Blob level (#238)
 - Windows support added (#257)
 - CI: build and test on Windows (#261)
@@ -14,6 +28,7 @@ Pearl changelog
 - Update offset only when write is imminent to prevent data corruption due to async Future cancellation (#228)
 - Improved Arc placement in Storage (#185)
 - Improved visibility levels in Storage (#186)
+- `Clone` impl removed from `Storage` (#265)
 - BLOB-wide exclusive lock removed from the write path, instead the short lock on the in-memory index added (#249)
 - Include BTreeMap overhead in index calculations (#266)
 - `MEM_SIZE` const has been added into the `Key` trait for proper calculation of memory occupied by the Key struct (#270)